[prod_env] simplify binary value order tests
